% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{add_group_component}
\alias{add_group_component}
\title{Adds a component with given name that contains all of the components
in group.}
\usage{
add_group_component(components, name, group)
}
\arguments{
\item{components}{Dataframe with components.}

\item{name}{Name of new group component.}

\item{group}{List of components that form the group.}
}
\value{
Dataframe with components.
}
\description{
Adds a component with given name that contains all of the components
in group.
}
\keyword{internal}
