% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_backends.R
\name{check_cmdstanr}
\alias{check_cmdstanr}
\title{Check that the required packages for using the cmdstanr backend are installed.}
\usage{
check_cmdstanr()
}
\value{
NULL if successful, and prints the current version of cmdstan being used.
}
\description{
Check that the required packages for using the cmdstanr backend are installed.
}
\keyword{internal}
