% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{compile_stan_model}
\alias{compile_stan_model}
\title{Compile Stan model}
\usage{
compile_stan_model(model)

compile_stan_model(model)

compile_stan_model(model)
}
\arguments{
\item{model}{String 'linear' or 'logistic' to specify a linear or logistic
trend.}

\item{model}{String 'linear' or 'logistic' to specify a linear or logistic
trend.}

\item{model}{String 'linear' or 'logistic' to specify a linear or logistic
trend.}
}
\value{
Stan model.

Stan model.

Stan model.
}
\description{
Compile Stan model

Compile Stan model

Compile Stan model
}
\keyword{internal}

