% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{compile_stan_model_ht}
\alias{compile_stan_model_ht}
\title{Compile Stan model for hierachical timeseries}
\usage{
compile_stan_model_ht(model)
}
\arguments{
\item{model}{String 'ht_linear' or "ht_linear_cov' to specify a linear 
model with or without considering covariance structure of growth trend}
}
\value{
Stan model.
}
\description{
Compile Stan model for hierachical timeseries
}
\note{
hard code the path, should change when making package
}

