% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{construct_holiday_dataframe}
\alias{construct_holiday_dataframe}
\title{Construct a dataframe of holiday dates.}
\usage{
construct_holiday_dataframe(m, dates)
}
\arguments{
\item{m}{Prophet object.}

\item{dates}{Vector with dates used for computing seasonality.}
}
\value{
A dataframe of holiday dates, in holiday dataframe format used in
 initialization.
}
\description{
Construct a dataframe of holiday dates.
}
\keyword{internal}
