% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{coverage}
\alias{coverage}
\title{Coverage}
\usage{
coverage(df, w)
}
\arguments{
\item{df}{Cross-validation results dataframe.}

\item{w}{Aggregation window size.}
}
\value{
Array of coverages
}
\description{
Coverage
}
\keyword{internal}
