% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{create_metric_data}
\alias{create_metric_data}
\title{Prepare dataframe for metrics calculation.}
\usage{
create_metric_data(m = NULL, df = NULL)
}
\arguments{
\item{m}{Prophet object. Default NULL}

\item{df}{A dataframe which is output of `simulated_historical_forecasts` or `cross_validation` Default NULL}
}
\value{
A dataframe only with y and yhat as a column.
}
\description{
Prepare dataframe for metrics calculation.
}
\keyword{internal}
