% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{df_for_plotting}
\alias{df_for_plotting}
\title{Merge history and forecast for plotting.}
\usage{
df_for_plotting(m, fcst)

df_for_plotting(m, fcst)

df_for_plotting(m, fcst)
}
\arguments{
\item{m}{Prophet object.}

\item{fcst}{Data frame returned by prophet predict.}

\item{m}{Prophet object.}

\item{fcst}{Data frame returned by prophet predict.}

\item{m}{Prophet object.}

\item{fcst}{Data frame returned by prophet predict.}
}
\description{
Merge history and forecast for plotting.

Merge history and forecast for plotting.

Merge history and forecast for plotting.
}
\keyword{internal}

