% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_backends.R
\name{.load_model}
\alias{.load_model}
\title{Load the Prophet Stan model.}
\usage{
.load_model(backend)
}
\arguments{
\item{backend}{"rstan" or "cmdstanr".}
}
\value{
stanmodel object if backend = "rstan", CmdStanModel object if backend = "cmdstanr"
}
\description{
Load the Prophet Stan model.
}
\keyword{internal}
