% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_backends.R
\name{.sampling}
\alias{.sampling}
\title{Obtain the joint posterior distribution of the parameters of the Prophet
model using MCMC sampling.}
\usage{
.sampling(args, backend)
}
\arguments{
\item{args}{Named list of arguments suitable for the chosen backend. Must
include arguments required for MCMC sampling.}

\item{backend}{"rstan" or "cmdstanr".}
}
\value{
A named list containing "stan_fit" (the fitted stan object),
  "params", and "n_iteration"
}
\description{
Obtain the joint posterior distribution of the parameters of the Prophet
model using MCMC sampling.
}
\keyword{internal}
