% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R
\name{fit_ht.prophet}
\alias{fit_ht.prophet}
\title{Fit the hierarchical time series prophet model.}
\usage{
fit_ht.prophet(m, df, ...)

fit_ht.prophet(m, df, ...)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Data frame.}

\item{...}{Additional arguments passed to the \code{optimizing} or
\code{sampling} functions in Stan.}

\item{m}{Prophet object.}

\item{df}{Data frame.}

\item{...}{Additional arguments passed to the \code{optimizing} or
\code{sampling} functions in Stan.}
}
\description{
This sets m$params to contain the fitted model parameters. It is a list
with the following elements:
  k (M array): M posterior samples of the initial slope.
  m (M array): The initial intercept.
  delta (MxN matrix): The slope change at each of N changepoints.
  beta (MxK matrix): Coefficients for K seasonality features.
  sigma_obs (M array): Noise level.
Note that M=1 if MAP estimation.

This sets m$params to contain the fitted model parameters. It is a list
with the following elements:
  k (M array): M posterior samples of the initial slope.
  m (M array): The initial intercept.
  delta (MxN matrix): The slope change at each of N changepoints.
  beta (MxK matrix): Coefficients for K seasonality features.
  sigma_obs (M array): Noise level.
Note that M=1 if MAP estimation.
}

