% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{flat_growth_init}
\alias{flat_growth_init}
\title{Initialize flat growth.}
\usage{
flat_growth_init(df)
}
\arguments{
\item{df}{Data frame with columns ds (date), y_scaled (scaled time series),
and t (scaled time).}
}
\value{
A vector (k, m) with the rate (k) and offset (m) of the flat
 growth function.
}
\description{
Provides a strong initialization for flat growth by setting the
growth to 0 and calculates the offset parameter that pass the
function through the mean of the the y_scaled values.
}
\keyword{internal}
