% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{flat_trend}
\alias{flat_trend}
\title{Evaluate the flat trend function.}
\usage{
flat_trend(t, m)
}
\arguments{
\item{t}{Vector of times on which the function is evaluated.}

\item{m}{Float initial offset.}
}
\value{
Vector y(t).
}
\description{
Evaluate the flat trend function.
}
\keyword{internal}
