% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{fourier_series}
\alias{fourier_series}
\title{Provides Fourier series components with the specified frequency and order.}
\usage{
fourier_series(dates, period, series.order)

fourier_series(dates, period, series.order)

fourier_series(dates, period, series.order)
}
\arguments{
\item{dates}{Vector of dates.}

\item{period}{Number of days of the period.}

\item{series.order}{Number of components.}

\item{dates}{Vector of dates.}

\item{period}{Number of days of the period.}

\item{series.order}{Number of components.}

\item{dates}{Vector of dates.}

\item{period}{Number of days of the period.}

\item{series.order}{Number of components.}
}
\value{
Matrix with seasonality features.

Matrix with seasonality features.

Matrix with seasonality features.
}
\description{
Provides Fourier series components with the specified frequency and order.

Provides Fourier series components with the specified frequency and order.

Provides Fourier series components with the specified frequency and order.
}
\keyword{internal}

