% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{generate_cutoffs}
\alias{generate_cutoffs}
\title{Generate cutoff dates}
\usage{
generate_cutoffs(df, horizon, initial, period)
}
\arguments{
\item{df}{Dataframe with historical data.}

\item{horizon}{timediff forecast horizon.}

\item{initial}{timediff initial window.}

\item{period}{timediff Simulated forecasts are done with this period.}
}
\value{
Array of datetimes.
}
\description{
Generate cutoff dates
}
\keyword{internal}
