% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generated_holidays}
\alias{generated_holidays}
\title{Generated table of holiday dates at the country level from 1995 to 2045}
\format{
A data frame with four variables: ds, holiday, country, year
}
\source{
\url{https://github.com/facebook/prophet/blob/main/python/scripts/generate_holidays_file.py}
}
\usage{
generated_holidays
}
\description{
The data is primarily based on the Python package [holidays](https://pypi.org/project/holidays/)
}
\keyword{datasets}
