% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{get_changepoint_indexes}
\alias{get_changepoint_indexes}
\title{Gets changepoint indexes in history dataframe.}
\usage{
get_changepoint_indexes(m)
}
\arguments{
\item{m}{Prophet object.}
}
\value{
array of indexes.
}
\description{
Gets changepoint indexes in history dataframe.
}

