% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{get_changepoint_times}
\alias{get_changepoint_times}
\title{Gets changepoint times, in scaled space.}
\usage{
get_changepoint_times(m)
}
\arguments{
\item{m}{Prophet object.}
}
\value{
array of times.
}
\description{
Gets changepoint times, in scaled space.
}

