% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_holidays.R
\name{get_holiday_names}
\alias{get_holiday_names}
\title{Return all possible holiday names of given country}
\usage{
get_holiday_names(country.name)
}
\arguments{
\item{country.name}{Country name (character).}
}
\value{
A vector of all possible holiday names (unique) of given country.
}
\description{
Return all possible holiday names of given country
}
\keyword{internal}
