% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_backends.R
\name{get_stan_backend}
\alias{get_stan_backend}
\title{Get the stan backend defined in the environment variables.}
\usage{
get_stan_backend()
}
\value{
'rstan' or 'cmdstanr'. 'rstan' if variable is not set.
}
\description{
Get the stan backend defined in the environment variables.
}
\keyword{internal}
