% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R
\name{ht_linear_growth_init}
\alias{ht_linear_growth_init}
\title{Initialize hierachical time series linear growth.}
\usage{
ht_linear_growth_init(df)

ht_linear_growth_init(df)
}
\arguments{
\item{df}{Data frame with columns ds (date), y_scaled (scaled time series),
and t (scaled time).}

\item{df}{Data frame with columns ds (date), y_scaled (scaled time series),
and t (scaled time).}
}
\value{
A vector (k, m) with the rate (k) and offset (m) of the linear
 growth function.

A vector (k, m) with the rate (k) and offset (m) of the linear
 growth function.
}
\description{
Provides a strong initialization for linear growth by calculating the
growth and offset parameters that pass the function through the first and
last points in the time series.

Provides a strong initialization for linear growth by calculating the
growth and offset parameters that pass the function through the first and
last points in the time series.
}

