% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{initialize_scales_fn}
\alias{initialize_scales_fn}
\title{Initialize model scales.}
\usage{
initialize_scales_fn(m, initialize_scales, df)
}
\arguments{
\item{m}{Prophet object.}

\item{initialize_scales}{Boolean set the scales or not.}

\item{df}{Dataframe for setting scales.}
}
\value{
Prophet object with scales set.
}
\description{
Sets model scaling factors using df.
}
\keyword{internal}
