% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{linear_growth_init}
\alias{linear_growth_init}
\title{Initialize linear growth.}
\usage{
linear_growth_init(df)
}
\arguments{
\item{df}{Data frame with columns ds (date), y_scaled (scaled time series),
and t (scaled time).}
}
\value{
A vector (k, m) with the rate (k) and offset (m) of the linear
 growth function.
}
\description{
Provides a strong initialization for linear growth by calculating the
growth and offset parameters that pass the function through the first and
last points in the time series.
}
\keyword{internal}
