% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{logistic_growth_init}
\alias{logistic_growth_init}
\title{Initialize logistic growth.}
\usage{
logistic_growth_init(df)
}
\arguments{
\item{df}{Data frame with columns ds (date), cap_scaled (scaled capacity),
y_scaled (scaled time series), and t (scaled time).}
}
\value{
A vector (k, m) with the rate (k) and offset (m) of the logistic
 growth function.
}
\description{
Provides a strong initialization for logistic growth by calculating the
growth and offset parameters that pass the function through the first and
last points in the time series.
}
\keyword{internal}
