% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(df, w)
}
\arguments{
\item{df}{Cross-validation results dataframe.}

\item{w}{Aggregation window size.}
}
\value{
Array of mean absolute errors.
}
\description{
Mean absolute error
}
\keyword{internal}
