% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{make_all_seasonality_features}
\alias{make_all_seasonality_features}
\title{Dataframe with seasonality features.}
\usage{
make_all_seasonality_features(m, df)

make_all_seasonality_features(m, df)

make_all_seasonality_features(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Dataframe with dates for computing seasonality features.}

\item{m}{Prophet object.}

\item{df}{Dataframe with dates for computing seasonality features.}

\item{m}{Prophet object.}

\item{df}{Dataframe with dates for computing seasonality features.}
}
\value{
Dataframe with seasonality.

Dataframe with seasonality.

Dataframe with seasonality.
}
\description{
Dataframe with seasonality features.

Dataframe with seasonality features.

Dataframe with seasonality features.
}
\keyword{internal}

