% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{make_holiday_features}
\alias{make_holiday_features}
\title{Construct a matrix of holiday features.}
\usage{
make_holiday_features(m, dates)
}
\arguments{
\item{m}{Prophet object.}

\item{dates}{Vector with dates used for computing seasonality.}
}
\value{
A dataframe with a column for each holiday
}
\description{
Construct a matrix of holiday features.
}

