% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_holidays.R
\name{make_holidays_df}
\alias{make_holidays_df}
\title{Make dataframe of holidays for given years and countries}
\usage{
make_holidays_df(years, country.name)
}
\arguments{
\item{years}{List of years for which to include holiday dates.}

\item{country.name}{Country name (character).}
}
\value{
Dataframe with 'ds' and 'holiday', which can directly feed
 to 'holidays' params in Prophet
}
\description{
Make dataframe of holidays for given years and countries
}
\keyword{internal}
