% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{make_metrics_function}
\alias{make_metrics_function}
\title{Meta function to make the function which evaluate metrics.}
\usage{
make_metrics_function(metrics)
}
\arguments{
\item{metrics}{metrics function}
}
\value{
A function using for metrics evaluation.
}
\description{
Meta function to make the function which evaluate metrics.
}
\keyword{internal}
