% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{mape}
\alias{mape}
\title{Mean absolute percent error}
\usage{
mape(df, w)
}
\arguments{
\item{df}{Cross-validation results dataframe.}

\item{w}{Aggregation window size.}
}
\value{
Array of mean absolute percent errors.
}
\description{
Mean absolute percent error
}
\keyword{internal}
