% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{mdape}
\alias{mdape}
\title{Median absolute percent error}
\usage{
mdape(df, w)
}
\arguments{
\item{df}{Cross-validation results dataframe.}

\item{w}{Aggregation window size.}
}
\value{
Array of median absolute percent errors.
}
\description{
Median absolute percent error
}
\keyword{internal}
