% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{mse}
\alias{mse}
\title{Mean squared error}
\usage{
mse(df, w)
}
\arguments{
\item{df}{Cross-validation results dataframe.}

\item{w}{Aggregation window size.}
}
\value{
Array of mean squared errors.
}
\description{
Mean squared error
}
\keyword{internal}
