% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{parse_seasonality_args}
\alias{parse_seasonality_args}
\title{Get number of Fourier components for built-in seasonalities.}
\usage{
parse_seasonality_args(m, name, arg, auto.disable, default.order)
}
\arguments{
\item{m}{Prophet object.}

\item{name}{String name of the seasonality component.}

\item{arg}{'auto', TRUE, FALSE, or number of Fourier components as
provided.}

\item{auto.disable}{Bool if seasonality should be disabled when 'auto'.}

\item{default.order}{Int default Fourier order.}
}
\value{
Number of Fourier components, or 0 for disabled.
}
\description{
Get number of Fourier components for built-in seasonalities.
}
\keyword{internal}
