% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{piecewise_linear}
\alias{piecewise_linear}
\title{Evaluate the piecewise linear function.}
\usage{
piecewise_linear(t, deltas, k, m, changepoint.ts)

piecewise_linear(t, deltas, k, m, changepoint.ts)

piecewise_linear(t, deltas, k, m, changepoint.ts)
}
\arguments{
\item{t}{Vector of times on which the function is evaluated.}

\item{deltas}{Vector of rate changes at each changepoint.}

\item{k}{Float initial rate.}

\item{m}{Float initial offset.}

\item{changepoint.ts}{Vector of changepoint times.}

\item{t}{Vector of times on which the function is evaluated.}

\item{deltas}{Vector of rate changes at each changepoint.}

\item{k}{Float initial rate.}

\item{m}{Float initial offset.}

\item{changepoint.ts}{Vector of changepoint times.}

\item{t}{Vector of times on which the function is evaluated.}

\item{deltas}{Vector of rate changes at each changepoint.}

\item{k}{Float initial rate.}

\item{m}{Float initial offset.}

\item{changepoint.ts}{Vector of changepoint times.}
}
\value{
Vector y(t).

Vector y(t).

Vector y(t).
}
\description{
Evaluate the piecewise linear function.

Evaluate the piecewise linear function.

Evaluate the piecewise linear function.
}
\keyword{internal}

