% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{piecewise_logistic}
\alias{piecewise_logistic}
\title{Evaluate the piecewise logistic function.}
\usage{
piecewise_logistic(t, cap, deltas, k, m, changepoint.ts)

piecewise_logistic(t, cap, deltas, k, m, changepoint.ts)

piecewise_logistic(t, cap, deltas, k, m, changepoint.ts)
}
\arguments{
\item{t}{Vector of times on which the function is evaluated.}

\item{cap}{Vector of capacities at each t.}

\item{deltas}{Vector of rate changes at each changepoint.}

\item{k}{Float initial rate.}

\item{m}{Float initial offset.}

\item{changepoint.ts}{Vector of changepoint times.}

\item{t}{Vector of times on which the function is evaluated.}

\item{cap}{Vector of capacities at each t.}

\item{deltas}{Vector of rate changes at each changepoint.}

\item{k}{Float initial rate.}

\item{m}{Float initial offset.}

\item{changepoint.ts}{Vector of changepoint times.}

\item{t}{Vector of times on which the function is evaluated.}

\item{cap}{Vector of capacities at each t.}

\item{deltas}{Vector of rate changes at each changepoint.}

\item{k}{Float initial rate.}

\item{m}{Float initial offset.}

\item{changepoint.ts}{Vector of changepoint times.}
}
\value{
Vector y(t).

Vector y(t).

Vector y(t).
}
\description{
Evaluate the piecewise logistic function.

Evaluate the piecewise logistic function.

Evaluate the piecewise logistic function.
}
\keyword{internal}

