% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_forecast_component}
\alias{plot_forecast_component}
\title{Plot a particular component of the forecast.}
\usage{
plot_forecast_component(m, fcst, name, uncertainty = TRUE, plot_cap = FALSE)
}
\arguments{
\item{m}{Prophet model}

\item{fcst}{Dataframe output of `predict`.}

\item{name}{String name of the component to plot (column of fcst).}

\item{uncertainty}{Boolean to plot uncertainty intervals.}

\item{plot_cap}{Boolean indicating if the capacity should be shown in the
figure, if available.}
}
\value{
A ggplot2 plot.
}
\description{
Plot a particular component of the forecast.
}
