% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet_cov.R
\name{predict_ht_seasonal_components}
\alias{predict_ht_seasonal_components}
\title{Predict seasonality broken down into components.}
\usage{
predict_ht_seasonal_components(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}
}
\value{
Dataframe with seasonal components.
}
\description{
Predict seasonality broken down into components.
}

