% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{predict_seasonal_components_ht}
\alias{predict_seasonal_components_ht}
\title{Predict seasonality broken down into components.}
\usage{
predict_seasonal_components_ht(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}
}
\value{
Dataframe with seasonal components.
}
\description{
Predict seasonality broken down into components.
}

