% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{predict_trend_ht}
\alias{predict_trend_ht}
\title{Predict trend using the prophet model.}
\usage{
predict_trend_ht(model, df)
}
\arguments{
\item{model}{Prophet object.}

\item{df}{Prediction dataframe.}
}
\value{
Vector with trend on prediction dates.
}
\description{
Predict trend using the prophet model.
}

