% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{predict_uncertainty}
\alias{predict_uncertainty}
\title{Prophet uncertainty intervals.}
\usage{
predict_uncertainty(m, df)

predict_uncertainty(m, df)

predict_uncertainty(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}

\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}

\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}
}
\value{
Dataframe with uncertainty intervals.

Dataframe with uncertainty intervals.

Dataframe with uncertainty intervals.
}
\description{
Prophet uncertainty intervals.

Prophet uncertainty intervals.

Prophet uncertainty intervals.
}
\keyword{internal}

