% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{predict_uncertainty_ht}
\alias{predict_uncertainty_ht}
\title{Prophet uncertainty intervals for hierachical time series}
\usage{
predict_uncertainty_ht(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}
}
\value{
Dataframe with uncertainty intervals.
}
\description{
Prophet uncertainty intervals for hierachical time series
}

