% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{predictive_samples}
\alias{predictive_samples}
\title{Sample from the posterior predictive distribution.}
\usage{
predictive_samples(m, df)

predictive_samples(m, df)

predictive_samples(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Dataframe with dates for predictions (column ds), and capacity
(column cap) if logistic growth.}

\item{m}{Prophet object.}

\item{df}{Dataframe with dates for predictions (column ds), and capacity
(column cap) if logistic growth.}

\item{m}{Prophet object.}

\item{df}{Dataframe with dates for predictions (column ds), and capacity
(column cap) if logistic growth.}
}
\value{
A list with items "trend", "seasonal", and "yhat" containing
 posterior predictive samples for that component.

A list with items "trend", "seasonal", and "yhat" containing
 posterior predictive samples for that component.

A list with items "trend", "seasonal", and "yhat" containing
 posterior predictive samples for that component.
}
\description{
Sample from the posterior predictive distribution.

Sample from the posterior predictive distribution.

Sample from the posterior predictive distribution.
}

