% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{prophet_copy}
\alias{prophet_copy}
\title{Copy Prophet object.}
\usage{
prophet_copy(m, cutoff = NULL)
}
\arguments{
\item{m}{Prophet model object.}

\item{cutoff}{Date, possibly as string. Changepoints are only retained if
changepoints <= cutoff.}
}
\value{
An unfitted Prophet model object with the same parameters as the
 input model.
}
\description{
Copy Prophet object.
}
\keyword{internal}
