% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{regressor_column_matrix}
\alias{regressor_column_matrix}
\title{Dataframe indicating which columns of the feature matrix correspond to
which seasonality/regressor components.}
\usage{
regressor_column_matrix(m, seasonal.features, modes)
}
\arguments{
\item{m}{Prophet object.}

\item{seasonal.features}{Constructed seasonal features dataframe.}

\item{modes}{List with keys 'additive' and 'multiplicative' with arrays of
component names for each mode of seasonality.}
}
\value{
List with items
 component.cols: A binary indicator dataframe with columns seasonal
   components and rows columns in seasonal.features. Entry is 1 if that
   column is used in that component.
 modes: Updated input with combination components.
}
\description{
Includes combination components, like 'additive_terms'. These combination
components will be added to the 'modes' input.
}
\keyword{internal}
