% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{rmse}
\alias{rmse}
\title{Root mean squared error}
\usage{
rmse(df, w)
}
\arguments{
\item{df}{Cross-validation results dataframe.}

\item{w}{Aggregation window size.}
}
\value{
Array of root mean squared errors.
}
\description{
Root mean squared error
}
\keyword{internal}
