% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{rolling_mean}
\alias{rolling_mean}
\title{Compute a rolling mean of x}
\usage{
rolling_mean(x, w)
}
\arguments{
\item{x}{Array.}

\item{w}{Integer window size (number of elements).}
}
\value{
Rolling mean of x with window size w.
}
\description{
Right-aligned. Padded with NAs on the front so the output is the same
size as x.
}
\keyword{internal}
