% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{rolling_mean_by_h}
\alias{rolling_mean_by_h}
\title{Compute a rolling mean of x, after first aggregating by h}
\usage{
rolling_mean_by_h(x, h, w, name)
}
\arguments{
\item{x}{Array.}

\item{h}{Array of horizon for each value in x.}

\item{w}{Integer window size (number of elements).}

\item{name}{String name for metric in result dataframe.}
}
\value{
Dataframe with columns horizon and name, the rolling mean of x.
}
\description{
Right-aligned. Computes a single mean for each unique value of h. Each mean
is over at least w samples.
}
\keyword{internal}
