% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{rolling_median_by_h}
\alias{rolling_median_by_h}
\title{Compute a rolling median of x, after first aggregating by h}
\usage{
rolling_median_by_h(x, h, w, name)
}
\arguments{
\item{x}{Array.}

\item{h}{Array of horizon for each value in x.}

\item{w}{Integer window size (number of elements).}

\item{name}{String name for metric in result dataframe.}
}
\value{
Dataframe with columns horizon and name, the rolling median of x.
}
\description{
Right-aligned. Computes a single median for each unique value of h. Each median
is over at least w samples.
}
\details{
For each h where there are fewer than w samples, we take samples from the previous h,
}
