% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{sample_model}
\alias{sample_model}
\title{Simulate observations from the extrapolated generative model.}
\usage{
sample_model(m, df, seasonal.features, iteration)

sample_model(m, df, seasonal.features, iteration)

sample_model(m, df, seasonal.features, iteration)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}

\item{seasonal.features}{Data frame of seasonal features}

\item{iteration}{Int sampling iteration to use parameters from.}

\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}

\item{seasonal.features}{Data frame of seasonal features}

\item{iteration}{Int sampling iteration to use parameters from.}

\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}

\item{seasonal.features}{Data frame of seasonal features}

\item{iteration}{Int sampling iteration to use parameters from.}
}
\value{
List of trend, seasonality, and yhat, each a vector like df$t.

List of trend, seasonality, and yhat, each a vector like df$t.

List of trend, seasonality, and yhat, each a vector like df$t.
}
\description{
Simulate observations from the extrapolated generative model.

Simulate observations from the extrapolated generative model.

Simulate observations from the extrapolated generative model.
}
\keyword{internal}

