% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{sample_posterior_predictive_ht}
\alias{sample_posterior_predictive_ht}
\title{Prophet posterior predictive samples for hierarchical time series.}
\usage{
sample_posterior_predictive_ht(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}
}
\value{
List with posterior predictive samples for each component.
}
\description{
Prophet posterior predictive samples for hierarchical time series.
}

