% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{sample_predictive_trend}
\alias{sample_predictive_trend}
\title{Simulate the trend using the extrapolated generative model.}
\usage{
sample_predictive_trend(model, df, iteration)

sample_predictive_trend(model, df, iteration)

sample_predictive_trend(model, df, iteration)
}
\arguments{
\item{model}{Prophet object.}

\item{df}{Prediction dataframe.}

\item{iteration}{Int sampling iteration to use parameters from.}

\item{model}{Prophet object.}

\item{df}{Prediction dataframe.}

\item{iteration}{Int sampling iteration to use parameters from.}

\item{model}{Prophet object.}

\item{df}{Prediction dataframe.}

\item{iteration}{Int sampling iteration to use parameters from.}
}
\value{
Vector of simulated trend over df$t.

Vector of simulated trend over df$t.

Vector of simulated trend over df$t.
}
\description{
Simulate the trend using the extrapolated generative model.

Simulate the trend using the extrapolated generative model.

Simulate the trend using the extrapolated generative model.
}
\keyword{internal}

