% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{sample_predictive_trend_ht}
\alias{sample_predictive_trend_ht}
\title{Simulate the trend using the extrapolated generative model for 
hierarchical time series.}
\usage{
sample_predictive_trend_ht(model, df)
}
\arguments{
\item{model}{Prophet object.}

\item{df}{Prediction dataframe.}

\item{iteration}{Int sampling iteration to use parameters from.}
}
\value{
Vector of simulated trend over df$t.
}
\description{
Simulate the trend using the extrapolated generative model for 
hierarchical time series.
}

