% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{seasonality_plot_df}
\alias{seasonality_plot_df}
\title{Prepare dataframe for plotting seasonal components.}
\usage{
seasonality_plot_df(m, ds)
}
\arguments{
\item{m}{Prophet object.}

\item{ds}{Array of dates for column ds.}
}
\value{
A dataframe with seasonal components on ds.
}
\description{
Prepare dataframe for plotting seasonal components.
}
\keyword{internal}
