% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{set_auto_seasonalities}
\alias{set_auto_seasonalities}
\title{Set seasonalities that were left on auto.}
\usage{
set_auto_seasonalities(m)
}
\arguments{
\item{m}{Prophet object.}
}
\value{
The prophet model with seasonalities set.
}
\description{
Turns on yearly seasonality if there is >=2 years of history.
Turns on weekly seasonality if there is >=2 weeks of history, and the
spacing between dates in the history is <7 days.
Turns on daily seasonality if there is >=2 days of history, and the spacing
between dates in the history is <1 day.
}
\keyword{internal}
