% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{setup_timezone}
\alias{setup_timezone}
\title{Sets the timezone environment variable to the timezone of the system.
Defaults to 'GMT' if there is no valid system timezone in the data or 
the system.}
\usage{
setup_timezone(m, df)
}
\arguments{
\item{m}{Prophet model.}

\item{df}{Dataframe containing the history.}
}
\description{
Sets the timezone environment variable to the timezone of the system.
Defaults to 'GMT' if there is no valid system timezone in the data or 
the system.
}
\keyword{internal}
