% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{single_cutoff_forecast}
\alias{single_cutoff_forecast}
\title{Forecast for a single cutoff.
Used in cross_validation function when evaluating for multiple cutoffs.}
\usage{
single_cutoff_forecast(df, model, cutoff, horizon.dt, predict_columns)
}
\arguments{
\item{df}{Dataframe with history for cutoff.}

\item{model}{Prophet model object.}

\item{cutoff}{Datetime of cutoff.}

\item{horizon.dt}{timediff forecast horizon.}

\item{predict_columns}{Array of names of columns to be returned in output.}
}
\value{
Dataframe with forecast, actual value, and cutoff.
}
\description{
Forecast for a single cutoff.
Used in cross_validation function when evaluating for multiple cutoffs.
}
\keyword{internal}
