% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{smape}
\alias{smape}
\title{Symmetric mean absolute percentage error
based on Chen and Yang (2004) formula}
\usage{
smape(df, w)
}
\arguments{
\item{df}{Cross-validation results dataframe.}

\item{w}{Aggregation window size.}
}
\value{
Array of symmetric mean absolute percent errors.
}
\description{
Symmetric mean absolute percentage error
based on Chen and Yang (2004) formula
}
\keyword{internal}
