% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{validate_column_name}
\alias{validate_column_name}
\title{Validates the name of a seasonality, holiday, or regressor.}
\usage{
validate_column_name(
  m,
  name,
  check_holidays = TRUE,
  check_seasonalities = TRUE,
  check_regressors = TRUE
)
}
\arguments{
\item{m}{Prophet object.}

\item{name}{string}

\item{check_holidays}{bool check if name already used for holiday}

\item{check_seasonalities}{bool check if name already used for seasonality}

\item{check_regressors}{bool check if name already used for regressor}
}
\description{
Validates the name of a seasonality, holiday, or regressor.
}
\keyword{internal}
