% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validates the inputs to Prophet.}
\usage{
validate_inputs(m)

validate_inputs(m)

validate_inputs(m)
}
\arguments{
\item{m}{Prophet object.}

\item{m}{Prophet object.}

\item{m}{Prophet object.}
}
\description{
Validates the inputs to Prophet.

Validates the inputs to Prophet.

Validates the inputs to Prophet.
}
\keyword{internal}

