% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{compile_stan_model}
\alias{compile_stan_model}
\title{Compile Stan model}
\usage{
compile_stan_model()
}
\arguments{
\item{model}{String 'linear' or 'logistic' to specify a linear or logistic
trend.}
}
\value{
Stan model.
}
\description{
Compile Stan model
}
\keyword{internal}
