% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{create_metric_data}
\alias{create_metric_data}
\title{Prepare dataframe for metrics calculation.}
\usage{
create_metric_data(m)
}
\arguments{
\item{m}{Prophet object.}
}
\value{
A dataframe only with y and yhat as a column.
}
\description{
Prepare dataframe for metrics calculation.
}
\keyword{internal}
