% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{fit.prophet}
\alias{fit.prophet}
\title{Fit the prophet model.}
\usage{
fit.prophet(m, df, ...)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Data frame.}

\item{...}{Additional arguments passed to the \code{optimizing} or 
\code{sampling} functions in Stan.}
}
\description{
Fit the prophet model.
}

