% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{fourier_series}
\alias{fourier_series}
\title{Provides fourier series components with the specified frequency.}
\usage{
fourier_series(dates, period, series.order)
}
\arguments{
\item{dates}{Vector of dates.}

\item{period}{Number of days of the period.}

\item{series.order}{Number of components.}
}
\value{
Matrix with seasonality features.
}
\description{
Provides fourier series components with the specified frequency.
}

