% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generated_holidays}
\alias{generated_holidays}
\title{holidays table}
\format{
A data frame with five variables: ds, holiday, country, year
}
\usage{
generated_holidays
}
\description{
holidays table
}
\keyword{datasets}
