% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{get_changepoint_matrix}
\alias{get_changepoint_matrix}
\title{Gets changepoint matrix for history dataframe.}
\usage{
get_changepoint_matrix(m)
}
\arguments{
\item{m}{Prophet object.}
}
\value{
array of indexes.
}
\description{
Gets changepoint matrix for history dataframe.
}
\keyword{internal}
