% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{get_prophet_stan_model}
\alias{get_prophet_stan_model}
\title{Load compiled Stan model}
\usage{
get_prophet_stan_model()
}
\arguments{
\item{model}{String 'linear' or 'logistic' to specify a linear or logistic
trend.}
}
\value{
Stan model.
}
\description{
Load compiled Stan model
}
\keyword{internal}
