% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{make_all_seasonality_features}
\alias{make_all_seasonality_features}
\title{Data frame seasonality features.}
\usage{
make_all_seasonality_features(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Dataframe with dates for computing seasonality features.}
}
\value{
Dataframe with seasonality.
}
\description{
Data frame seasonality features.
}

