% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{make_all_seasonality_features}
\alias{make_all_seasonality_features}
\title{Dataframe with seasonality features.
Includes seasonality features, holiday features, and added regressors.}
\usage{
make_all_seasonality_features(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Dataframe with dates for computing seasonality features and any
added regressors.}
}
\value{
List with items
 seasonal.features: Dataframe with regressor features,
 prior.scales: Array of prior scales for each colum of the features
   dataframe.
}
\description{
Dataframe with seasonality features.
Includes seasonality features, holiday features, and added regressors.
}
\keyword{internal}
