% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{make_future_dataframe}
\alias{make_future_dataframe}
\title{Make dataframe with future dates for forecasting.}
\usage{
make_future_dataframe(m, periods, freq = "d", include_history = TRUE)
}
\arguments{
\item{m}{Prophet model object.}

\item{periods}{Int number of periods to forecast forward.}

\item{freq}{'day', 'week', 'month', 'quarter', or 'year'.}

\item{include_history}{Boolean to include the historical dates in the data
frame for predictions.}
}
\value{
Dataframe that extends forward from the end of m$history for the
 requested number of periods.
}
\description{
Make dataframe with future dates for forecasting.
}

