% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{make_seasonality_features}
\alias{make_seasonality_features}
\title{Data frame with seasonality features.}
\usage{
make_seasonality_features(dates, period, series.order, prefix)
}
\arguments{
\item{dates}{Vector of dates.}

\item{period}{Number of days of the period.}

\item{series.order}{Number of components.}

\item{prefix}{Column name prefix.}
}
\value{
Dataframe with seasonality.
}
\description{
Data frame with seasonality features.
}
\keyword{internal}
