% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_seasonality}
\alias{plot_seasonality}
\title{Plot a custom seasonal component.}
\usage{
plot_seasonality(m, name, uncertainty = TRUE)
}
\arguments{
\item{m}{Prophet model object.}

\item{name}{String name of the seasonality.}

\item{uncertainty}{Boolean to plot uncertainty intervals.}
}
\value{
A ggplot2 plot.
}
\description{
Plot a custom seasonal component.
}
\keyword{internal}
