% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot the prophet trend.}
\usage{
plot_trend(df, uncertainty = TRUE, plot_cap = TRUE)

plot_trend(df, uncertainty = TRUE, plot_cap = TRUE)

plot_trend(df, uncertainty = TRUE, plot_cap = TRUE)
}
\arguments{
\item{df}{Forecast dataframe for plotting.}

\item{uncertainty}{Boolean to plot uncertainty intervals.}

\item{plot_cap}{Boolean indicating if the capacity should be shown in the
figure, if available.}

\item{df}{Forecast dataframe for plotting.}

\item{uncertainty}{Boolean to plot uncertainty intervals.}

\item{plot_cap}{Boolean indicating if the capacity should be shown in the
figure, if available.}

\item{df}{Forecast dataframe for plotting.}

\item{uncertainty}{Boolean to plot uncertainty intervals.}

\item{plot_cap}{Boolean indicating if the capacity should be shown in the
figure, if available.}
}
\value{
A ggplot2 plot.

A ggplot2 plot.

A ggplot2 plot.
}
\description{
Plot the prophet trend.

Plot the prophet trend.

Plot the prophet trend.
}
\keyword{internal}

