% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{plot_weekly}
\alias{plot_weekly}
\title{Plot the weekly component of the forecast.}
\usage{
plot_weekly(df, uncertainty = TRUE)
}
\arguments{
\item{df}{Forecast dataframe for plotting.}

\item{uncertainty}{Boolean to plot uncertainty intervals.}
}
\value{
A ggplot2 plot.
}
\description{
Plot the weekly component of the forecast.
}

