% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{plot_yearly}
\alias{plot_yearly}
\title{Plot the yearly component of the forecast.}
\usage{
plot_yearly(df, uncertainty = TRUE)
}
\arguments{
\item{df}{Forecast dataframe for plotting.}

\item{uncertainty}{Boolean to plot uncertainty intervals.}
}
\value{
A ggplot2 plot.
}
\description{
Plot the yearly component of the forecast.
}

