% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{predict_seasonal_components}
\alias{predict_seasonal_components}
\title{Seasonality broken down into components}
\usage{
predict_seasonal_components(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Data frame.}
}
\description{
Seasonality broken down into components
}

