% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{predict_seasonal_components}
\alias{predict_seasonal_components}
\title{Predict seasonality components, holidays, and added regressors.}
\usage{
predict_seasonal_components(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}
}
\value{
Dataframe with seasonal components.
}
\description{
Predict seasonality components, holidays, and added regressors.
}
\keyword{internal}
