% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R, R/prophet_cov.R, R/prophet_tmp.R
\name{predict_seasonal_components}
\alias{predict_seasonal_components}
\title{Predict seasonality broken down into components.}
\usage{
predict_seasonal_components(m, df)

predict_seasonal_components(m, df)

predict_seasonal_components(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}

\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}

\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}
}
\value{
Dataframe with seasonal components.

Dataframe with seasonal components.

Dataframe with seasonal components.
}
\description{
Predict seasonality broken down into components.

Predict seasonality broken down into components.

Predict seasonality broken down into components.
}
\keyword{internal}

