% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{predict_uncertainty}
\alias{predict_uncertainty}
\title{Prophet uncertainty intervals.}
\usage{
predict_uncertainty(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Data frame.}
}
\description{
Prophet uncertainty intervals.
}

