% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{predict_uncertainty}
\alias{predict_uncertainty}
\title{Prophet uncertainty intervals for yhat and trend}
\usage{
predict_uncertainty(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}
}
\value{
Dataframe with uncertainty intervals.
}
\description{
Prophet uncertainty intervals for yhat and trend
}
\keyword{internal}
