% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{prophet_plot_components}
\alias{prophet_plot_components}
\title{Plot the components of a prophet forecast.
Prints a ggplot2 with panels for trend, weekly and yearly seasonalities if
present, and holidays if present.}
\usage{
prophet_plot_components(m, fcst, uncertainty = TRUE)
}
\arguments{
\item{m}{Prophet object.}

\item{fcst}{Data frame returned by predict(m, df).}

\item{uncertainty}{Boolean indicating if the uncertainty interval should be
plotted for the trend, from fcst columns trend_lower and trend_upper.}
}
\value{
Invisibly return a list containing the plotted ggplot objects
}
\description{
Plot the components of a prophet forecast.
Prints a ggplot2 with panels for trend, weekly and yearly seasonalities if
present, and holidays if present.
}

