% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{sample_posterior_predictive}
\alias{sample_posterior_predictive}
\title{Prophet posterior predictive samples.}
\usage{
sample_posterior_predictive(m, df)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Prediction dataframe.}
}
\value{
List with posterior predictive samples for the forecast yhat and
 for the trend component.
}
\description{
Prophet posterior predictive samples.
}
\keyword{internal}
