% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{set_changepoints}
\alias{set_changepoints}
\title{Set changepoints}
\usage{
set_changepoints(m)
}
\arguments{
\item{m}{Prophet object.}
}
\value{
m with changepoints set.
}
\description{
Sets m$changepoints to the dates of changepoints.
}

