% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{set_changepoints}
\alias{set_changepoints}
\title{Set changepoints}
\usage{
set_changepoints(m)
}
\arguments{
\item{m}{Prophet object.}
}
\value{
m with changepoints set.
}
\description{
Sets m$changepoints to the dates of changepoints. Either:
1) The changepoints were passed in explicitly.
  A) They are empty.
  B) They are not empty, and need validation.
2) We are generating a grid of them.
3) The user prefers no changepoints be used.
}
\keyword{internal}
