% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{set_date}
\alias{set_date}
\title{Convert date vector}
\usage{
set_date(ds)
}
\arguments{
\item{ds}{Date vector}
}
\value{
vector of POSIXct object converted from date
}
\description{
Convert the date to POSIXct object. Timezones are stripped and replaced
with GMT.
}
\keyword{internal}
