% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{setup_dataframe}
\alias{setup_dataframe}
\title{Prepare dataframe for fitting or predicting.}
\usage{
setup_dataframe(m, df, initialize_scales = FALSE)
}
\arguments{
\item{m}{Prophet object.}

\item{df}{Data frame with columns ds, y, and cap if logistic growth. Any
specified additional regressors must also be present.}

\item{initialize_scales}{Boolean set scaling factors in m from df.}
}
\value{
list with items 'df' and 'm'.
}
\description{
Adds a time index and scales y. Creates auxillary columns 't', 't_ix',
'y_scaled', and 'cap_scaled'. These columns are used during both fitting
and predicting.
}
\keyword{internal}
