% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{time_diff}
\alias{time_diff}
\title{Time difference between datetimes}
\usage{
time_diff(ds1, ds2, units = "days")
}
\arguments{
\item{ds1}{POSIXct object}

\item{ds2}{POSIXct object}

\item{units}{string units of difference, e.g. 'days' or 'secs'.}
}
\value{
numeric time difference
}
\description{
Compute time difference of two POSIXct objects
}
\keyword{internal}
